# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CreateRechargeOrderBody": {
				"properties": {
					"packageId": {
						"type": "string"
					}
				},
				"required": [
					"packageId"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"LoginBody": {
				"description": "登录请求体",
				"properties": {
					"username": {
						"type": "string",
						"description": "用户名"
					},
					"password": {
						"type": "string",
						"description": "密码"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SendCodeBody": {
				"description": "发送验证码请求体",
				"properties": {
					"email": {
						"type": "string",
						"description": "邮箱地址"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RegisterBody": {
				"description": "注册请求体",
				"properties": {
					"username": {
						"type": "string",
						"description": "用户名"
					},
					"email": {
						"type": "string",
						"description": "邮箱"
					},
					"password": {
						"type": "string",
						"description": "密码（至少6位）"
					},
					"code": {
						"type": "string",
						"description": "邮箱验证码"
					}
				},
				"required": [
					"username",
					"email",
					"password",
					"code"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateModelBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"model_key": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"aspect_ratio": {
						"type": "string"
					},
					"aspect_ratios": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"consume_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"consume_points"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateModelBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"model_key": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"aspect_ratio": {
						"type": "string"
					},
					"aspect_ratios": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"consume_points": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "普通用户认证 Token"
			},
			"adminJwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "管理员认证 Token"
			}
		}
	},
	"info": {
		"title": "my-project",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
...
		"/app/recharge/orders": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"积分充值"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRechargeOrderBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"积分充值"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/app/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "用户登录\nPOST /app/login",
				"tags": [
					"用户认证"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBody"
							}
						}
					}
				}
			}
		},
		"/app/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "获取当前用户信息\nGET /app/me",
				"tags": [
					"用户认证"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/app/send-code": {
			"post": {
				"operationId": "SendCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "发送邮箱验证码\nPOST /app/send-code",
				"tags": [
					"用户认证"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendCodeBody"
							}
						}
					}
				}
			}
		},
		"/app/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "用户注册\nPOST /app/register",
				"tags": [
					"用户认证"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterBody"
							}
						}
					}
				}
			}
		},
		"/app/models": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "获取模型列表\nGET /app/models",
				"tags": [
					"模型"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "manufacturer",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/admin/orders": {
			"get": {
				"operationId": "ListOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"订单管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "orderNo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/admin/orders/{id}": {
			"get": {
				"operationId": "OrderDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"订单管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/admin/customers": {
			"get": {
				"operationId": "ListCustomers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"订单管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/admin/customers/{id}": {
			"get": {
				"operationId": "CustomerDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"订单管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/admin/customers/{id}/points-logs": {
			"get": {
				"operationId": "CustomerPointsLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"订单管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "changeType",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/admin/models": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateModelBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/admin/models/{id}": {
			"get": {
				"operationId": "Detail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateModelBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}